requirements/common.txt
Flask
Flask-SQLAlchemy
Flask-Script
Flask-Migrate
Flask-Admin
Flask-Login
Flask-WTF
Flask-Gears
SQLAlchemy
alembic
WTForms
Flask-Babel
Flask-Babelex
Flask-Mail
python-dateutil

Gears
gears-clean-css
gears-coffeescript
gears-less
gears-sass
gears-stylus
gears-uglifyjs

celery
git+https://github.com/SilentSokolov/flask-thumbnails